JSON Serialization
Swapped out the old and terrible string parsing system for a dedicated JSON serializer/deserializer. Using luajson as the lua-side json manager, and Newtonsoft on the C# side.

This should also open the door for an easy swap to protobuf (there is a protobuf lua package),  and also make it much easier to write lua api.